refactor(core)!: remove the dead passive scheduler abstraction
With the real WorkScheduler seam in place, the moirai-core::scheduler
module's passive abstraction is fully redundant and had no production
consumer. ThreadScheduler is active (its workers pop and execute tasks
internally), so the trait's next_task/steal_task hand-out contract could
only ever return a mock None on it.

Remove the Scheduler trait, the ScheduledTask erased-task type (a duplicate
of the executor's live ScheduledJob) and INLINE_SCHEDULED_TASK_WORDS, and
the SchedulerConfig/WorkStealingStrategy/QueueType/StealContext/Stats config
vocabulary that only served it. Collapse the module to its one live export,
SchedulerId (used by metrics aggregation). Update the core re-exports and
prelude, repoint the public-scheduler-task-surface contract and the
rayon/tokio gap audit at the executor's ScheduledJob (whose inline,
no-boxing erasure is guarded in source_contracts.rs), and prune the deleted
files from the contract module alias.

This leaves exactly one scheduler abstraction (the WorkScheduler seam) and
one erased-task type (ScheduledJob) -- no parallel dead duplicate.

BREAKING CHANGE: removed moirai_core::scheduler::{Scheduler, ScheduledTask,
INLINE_SCHEDULED_TASK_WORDS, SchedulerConfig, Config, WorkStealingStrategy,
QueueType, StealContext, Stats} and the corresponding re-exports/prelude
entries. SchedulerId is retained.
