+++ benchmarks/tests/benchmark_contracts/support.rs
@@ -75,9 +75,6 @@ const SPLIT_MODULE_ALIASES: &[(&str, &[&str])] = &[
         "../moirai-core/src/scheduler.rs",
         &[
             "../moirai-core/src/scheduler/mod.rs",
-            "../moirai-core/src/scheduler/traits.rs",
-            "../moirai-core/src/scheduler/config.rs",
-            "../moirai-core/src/scheduler/task.rs",
         ],
     ),
     (
