+++ docs/rayon_tokio_gap_audit.md
@@ -29,8 +29,8 @@ This audit covers the active unified-scheduler comparison scope:
 - PAL reactor task-handle completion: `IoReactor::spawn` publishes completion through per-task atomic state and wakes the awaiting `TaskHandle`.
 - PAL reactor task queue avoids dynamic future dispatch: queued reactor futures use bounded inline storage plus monomorphized poll/drop functions instead of `Pin<Box<dyn Future<Output = ()>>>`; oversized futures use a typed boxed fallback stored behind the same monomorphized path.
 - PAL platform reactor dispatch is static: `create_reactor` returns the compile-target `PlatformReactor` type instead of `Box<dyn Reactor>`.
-- Public scheduler task dispatch is concrete: `moirai-core::ScheduledTask` uses `INLINE_SCHEDULED_TASK_WORDS` inline storage with monomorphized execute/drop/context functions, and the canonical runtime scheduler `ThreadScheduler` queues monomorphized `ScheduledJob` values instead of `Box<dyn BoxedTask>`. (The redundant `WorkStealingScheduler`/`NumaAwareScheduler` were consolidated into `ThreadScheduler`; `moirai-scheduler` now provides only deque and NUMA-topology primitives.)
-- Standalone scheduler work-stealing storage is contiguous: `moirai-scheduler::ChaseLevDeque` stores `ScheduledTask` values in `UnsafeCell<MaybeUninit<T>>` ring slots instead of allocating one `Box<T>` per queued item. Reclamation is monomorphized by policy: default `QuiescentReclaim` uses zero-sized state and guard; opt-in `SharedEpochReclaim` uses one active-access counter to block shared reclamation while push, pop, or steal may hold an array pointer.
+- Public scheduler task dispatch is concrete: the canonical runtime scheduler `ThreadScheduler` queues monomorphized `ScheduledJob` values — inline `INLINE_JOB_WORDS` storage with monomorphized execute/drop and a typed `Box<T>` fallback only for oversized closures — instead of `Box<dyn BoxedTask>`. (The redundant `WorkStealingScheduler`/`NumaAwareScheduler`, and the dead `moirai-core` `Scheduler` trait + `ScheduledTask` duplicates, were removed; `moirai-scheduler` now provides only deque and NUMA-topology primitives, and the canonical scheduler abstraction is the executor's `WorkScheduler` seam.)
+- Standalone scheduler work-stealing storage is contiguous: `moirai-scheduler::ChaseLevDeque` stores its element values in `UnsafeCell<MaybeUninit<T>>` ring slots instead of allocating one `Box<T>` per queued item. Reclamation is monomorphized by policy: default `QuiescentReclaim` uses zero-sized state and guard; opt-in `SharedEpochReclaim` uses one active-access counter to block shared reclamation while push, pop, or steal may hold an array pointer.
 - Standalone deque reclamation policy overhead is measured by `thread_schedule_comparison -- standalone_deque_reclaim_policy`; the diagnostic keeps `QuiescentReclaim` as the production default and `SharedEpochReclaim` as an explicit opt-in policy.
 - PAL epoll reactor wake: Linux `EpollReactor::wake` uses an `eventfd` registered with epoll instead of a no-op wake path.
 - Async executor queueing: heterogeneous queued futures use monomorphized poll/drop function pointers instead of `dyn Future`.
@@ -183,7 +183,7 @@ Rayon does not expose a per-task result handle equivalent to `TaskHandle<T>`, so
 | PAL reactor task handles complete | spawned reactor tasks own per-task `TaskCompletion` state; completed tasks publish with release ordering and `TaskHandle` polls with acquire ordering | `spawned_ready_task_handle_completes_after_iteration` and `pal_async_io_facades_have_value_tests_and_self_wake_contract` |
 | PAL reactor task queue avoids dynamic future dispatch | `ErasedReactorTaskFuture` stores fitting futures inline in `ReactorTaskFutureStorage`, falls back to `Box<F>` only when the concrete future exceeds the static storage budget, and uses monomorphized poll/drop functions for both paths | `pal_async_io_facades_have_value_tests_and_self_wake_contract` |
 | PAL platform reactor dispatch is static | `PlatformReactor` is selected by target `cfg`, and `create_reactor` returns the concrete target type | `pal_async_io_facades_have_value_tests_and_self_wake_contract` |
-| Public scheduler task queues avoid dynamic task dispatch | `ScheduledTask` stores fitting tasks inline, falls back to typed `Box<T>` only for oversized tasks, and the canonical `ThreadScheduler` queues monomorphized `ScheduledJob` values | `public_scheduler_task_surface_uses_scheduled_task_erasure` |
+| Public scheduler task queues avoid dynamic task dispatch | The executor's `ScheduledJob` stores fitting tasks inline, falls back to typed `Box<T>` only for oversized tasks, and the canonical `ThreadScheduler` queues these monomorphized values | `public_scheduler_task_surface_uses_scheduled_task_erasure` |
 | Standalone scheduler deque avoids per-item heap nodes | `ChaseLevDeque<T, P>` writes tasks into contiguous `MaybeUninit<T>` ring slots and selects retired backing-array reclamation through a sealed policy. `QuiescentReclaim` has zero-sized state and guard; `SharedEpochReclaim` has one active-access counter and monomorphized guards. | `public_scheduler_task_surface_uses_scheduled_task_erasure`, `chase_lev_deque_*` unit tests, and the `standalone_deque_reclaim_policy` benchmark |
 | PAL epoll wake is real | Linux epoll registers an internal `eventfd`, `wake` writes to it, and `poll_events` drains it without emitting a user event | `pal_async_io_facades_have_value_tests_and_self_wake_contract` |
 | Iterator thread-pool jobs avoid dynamic dispatch | `ErasedThreadJob` owns a concrete job allocation and stores monomorphized run/drop functions | `iter_thread_pool_uses_monomorphized_erased_jobs` |
@@ -265,7 +265,7 @@ Rayon does not expose a per-task result handle equivalent to `TaskHandle<T>`, so
 - PAL reactor task handles must return `Ready(())` after the spawned task completes; tests must reject placeholder pending-only handles.
 - PAL reactor task queues must not use `Pin<Box<dyn Future<Output = ()>>>`; heterogeneous queued futures must use bounded inline storage and monomorphized poll/drop function pointers, with typed boxed fallback only for oversized futures.
 - PAL platform reactor selection must be static for the compile target and must not return `Box<dyn Reactor>`.
-- Public scheduler task APIs must not expose `Box<dyn BoxedTask>`, `dyn Scheduler`, `TaskSlot`, or `Pin<Box<dyn Future>>`; heterogeneous scheduler work must flow through `ScheduledTask`.
+- Public scheduler task APIs must not expose `Box<dyn BoxedTask>`, `dyn Scheduler`, `TaskSlot`, or `Pin<Box<dyn Future>>`; heterogeneous scheduler work must flow through the executor's inline-erased `ScheduledJob`.
 - Standalone scheduler deque storage must not reintroduce `Box<[AtomicPtr<T>]>`, `Box::into_raw(Box::new(item))`, or `Box::from_raw(item_ptr)` per-item queue nodes.
 - Standalone deque reclamation-policy diagnostics must assert the drained sum for quiescent and shared epoch policy rows before timing.
 - Work-class routing remains sealed, zero-sized, and monomorphized through `WorkClass` marker types.
